smartcareer-backend/requirements.txt
flask>=2.0.2
google-generativeai>=0.3.1
mysql-connector-python>=8.0.27
python-dotenv>=0.19.2
requests>=2.26.0
cachetools>=5.5.2